Stan data laag cli

+++ BezoekersRegistratieSysteem/BezoekersRegistratieSysteemREST/Model/Output/AfspraakOutputDTO.cs
@@ -1,20 +1,26 @@
 ﻿using BezoekersRegistratieSysteemBL.Domeinen;
 
-namespace BezoekersRegistratieSysteemREST.Model.Output {
-	public class AfsrpaakOutputDTO {
-		public static AfsrpaakOutputDTO NaarDTO(Afspraak afspraak) {
+namespace BezoekersRegistratieSysteemREST.Model.Output
+{
+	public class AfsrpaakOutputDTO
+	{
+		public static AfsrpaakOutputDTO NaarDTO(Afspraak afspraak)
+		{
 			return new(afspraak.Id, afspraak.Starttijd, afspraak.Eindtijd, afspraak.Bezoeker.Id, afspraak.Werknemer.Id);
 		}
 
-		public static IEnumerable<AfsrpaakOutputDTO> NaarDTO(IEnumerable<Afspraak> afspraken) {
+		public static IEnumerable<AfsrpaakOutputDTO> NaarDTO(IEnumerable<Afspraak> afspraken)
+		{
 			List<AfsrpaakOutputDTO> output = new();
-			foreach (Afspraak afspraak in afspraken) {
+			foreach (Afspraak afspraak in afspraken)
+			{
 				output.Add(AfsrpaakOutputDTO.NaarDTO(afspraak));
 			}
 			return output;
 		}
 
-		public AfsrpaakOutputDTO(uint id, DateTime starttijd, DateTime? eindtijd, uint bezoekerId, uint werknemerId) {
+		public AfsrpaakOutputDTO(long id, DateTime starttijd, DateTime? eindtijd, long bezoekerId, long werknemerId)
+		{
 			Id = id;
 			Starttijd = starttijd;
 			Eindtijd = eindtijd;
@@ -22,10 +28,10 @@ namespace BezoekersRegistratieSysteemREST.Model.Output {
 			WerknemerId = werknemerId;
 		}
 
-		public uint Id { get; set; }
+		public long Id { get; set; }
 		public DateTime Starttijd { get; set; }
 		public DateTime? Eindtijd { get; set; }
-		public uint BezoekerId { get; set; }
-		public uint WerknemerId { get; set; }
+		public long BezoekerId { get; set; }
+		public long WerknemerId { get; set; }
 	}
 }
