+++ BezoekersRegistratieSysteem/xUnitBezoekersRegistratiesysteem/Domeinen/UnitTestBedrijf.cs
@@ -1,8 +1,10 @@
 ﻿using BezoekersRegistratieSysteemBL.Domeinen;
 using BezoekersRegistratieSysteemBL.Exceptions.DomeinException;
 
-namespace xUnitBezoekersRegistratiesysteem.Domeinen {
-	public class UnitTestBedrijf {
+namespace xUnitBezoekersRegistratiesysteem.Domeinen
+{
+	public class UnitTestBedrijf
+	{
 		//AF
 
 		#region Valid Info
@@ -14,14 +16,16 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 
 		#region UnitTest Id
 		[Fact]
-		public void ZetId_Valid() {
+		public void ZetId_Valid()
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.ZetId(10);
-			Assert.Equal((uint)10, b.Id);
+			Assert.Equal((long)10, b.Id);
 		}
 
 		[Fact]
-		public void ZetId_Invalid() {
+		public void ZetId_Invalid()
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Assert.Throws<BedrijfException>(() => b.ZetId(0));
 		}
@@ -32,7 +36,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("bedrijf", "bedrijf")]
 		[InlineData("     bedrijf", "bedrijf")]
 		[InlineData("bedrijf     ", "bedrijf")]
-		public void ZetNaam_Valid(string naamIn, string naamUit) {
+		public void ZetNaam_Valid(string naamIn, string naamUit)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.ZetNaam(naamIn);
 			Assert.Equal(naamUit, b.Naam);
@@ -46,7 +51,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("\r")]
 		[InlineData("\t")]
 		[InlineData("\v")]
-		public void ZetNaam_Invalid(string naam) {
+		public void ZetNaam_Invalid(string naam)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Assert.Throws<BedrijfException>(() => b.ZetNaam(naam));
 		}
@@ -57,7 +63,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("BE0676747521", "BE0676747521")]
 		[InlineData("     BE0676747521", "BE0676747521")]
 		[InlineData("BE0676747521     ", "BE0676747521")]
-		public void ZetBTW_Valid(string btwIn, string btwUit) {
+		public void ZetBTW_Valid(string btwIn, string btwUit)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.ZetBTW(btwIn);
 			Assert.Equal(btwUit, b.BTW);
@@ -71,7 +78,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("\r")]
 		[InlineData("\t")]
 		[InlineData("\v")]
-		public void ZetBTW_Invalid(string btw) {
+		public void ZetBTW_Invalid(string btw)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Assert.Throws<BedrijfException>(() => b.ZetBTW(btw));
 		}
@@ -82,7 +90,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("BE0676747521", "BE0676747521")]
 		[InlineData("     BE0676747521", "BE0676747521")]
 		[InlineData("BE0676747521     ", "BE0676747521")]
-		public void ZetBTWControle_Valid(string btwIn, string btwUit) {
+		public void ZetBTWControle_Valid(string btwIn, string btwUit)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.ZetBTWControle(btwIn);
 			Assert.Equal(btwUit, b.BTW);
@@ -97,7 +106,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("\t")]
 		[InlineData("\v")]
 		[InlineData("BE123456789")] //nep BTWnummer
-		public void ZetBTWControle_Invalid(string btw) {
+		public void ZetBTWControle_Invalid(string btw)
+		{
 			Assert.Throws<BedrijfException>(() => new Bedrijf(10, "bedrijf", btw, "012345678", "bedrijf@email.com", "bedrijfstraat 10"));
 		}
 		#endregion
@@ -107,7 +117,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("012345678", "012345678")]
 		[InlineData("     012345678", "012345678")]
 		[InlineData("012345678     ", "012345678")]
-		public void ZetTelefoonnummer_Valid(string telefoonnummerIn, string telefoonnummerUit) {
+		public void ZetTelefoonnummer_Valid(string telefoonnummerIn, string telefoonnummerUit)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.ZetTelefoonNummer(telefoonnummerIn);
 			Assert.Equal(telefoonnummerUit, b.TelefoonNummer);
@@ -122,7 +133,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("\t")]
 		[InlineData("\v")]
 		[InlineData("0123456789101112")]
-		public void ZetTelefoonnummer_Invalid(string telefoonnummer) {
+		public void ZetTelefoonnummer_Invalid(string telefoonnummer)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Assert.Throws<BedrijfException>(() => b.ZetTelefoonNummer(telefoonnummer));
 		}
@@ -133,7 +145,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("bedrijf@email.com", "bedrijf@email.com")]
 		[InlineData("     bedrijf@email.com", "bedrijf@email.com")]
 		[InlineData("bedrijf@email.com     ", "bedrijf@email.com")]
-		public void ZetEmail_Valid(string emailIn, string emailUit) {
+		public void ZetEmail_Valid(string emailIn, string emailUit)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.ZetEmail(emailIn);
 			Assert.Equal(emailUit, b.Email);
@@ -155,7 +168,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("bedrijf")]
 		[InlineData("bedrijf@.")]
 		[InlineData("bedrijf.com")]
-		public void ZetEmail_Invalid(string email) {
+		public void ZetEmail_Invalid(string email)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Assert.Throws<BedrijfException>(() => b.ZetEmail(email));
 		}
@@ -166,7 +180,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("bedrijfstraat 10", "bedrijfstraat 10")]
 		[InlineData("     bedrijfstraat 10", "bedrijfstraat 10")]
 		[InlineData("bedrijfstraat 10     ", "bedrijfstraat 10")]
-		public void ZetAdres_Valid(string adresIn, string adresUit) {
+		public void ZetAdres_Valid(string adresIn, string adresUit)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.ZetAdres(adresIn);
 			Assert.Equal(adresUit, b.Adres);
@@ -180,7 +195,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("\r")]
 		[InlineData("\t")]
 		[InlineData("\v")]
-		public void ZetAdres_Invalid(string adres) {
+		public void ZetAdres_Invalid(string adres)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Assert.Throws<BedrijfException>(() => b.ZetAdres(adres));
 
@@ -189,7 +205,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 
 		#region UnitTest Voeg Werknemer Toe
 		[Fact]
-		public void VoegWerknemerToeInBedrijf_Valid() {
+		public void VoegWerknemerToeInBedrijf_Valid()
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.VoegWerknemerToeInBedrijf(_w, _e, _f1);
 			Assert.Contains(_w, b.GeefWerknemers());
@@ -221,14 +238,18 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData("\r", "functie")]
 		[InlineData("\t", "functie")]
 		[InlineData("\v", "functie")]
-		public void VoegWerknemerToeInBedrijf_Invalid(string email, string functie) {
+		public void VoegWerknemerToeInBedrijf_Invalid(string email, string functie)
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 
 			Assert.Throws<BedrijfException>(() => b.VoegWerknemerToeInBedrijf(null, _e, _f1));
 			Assert.Throws<BedrijfException>(() => b.VoegWerknemerToeInBedrijf(null, null, null));
-			if (string.IsNullOrWhiteSpace(functie)) Assert.Throws<BedrijfException>(() => b.VoegWerknemerToeInBedrijf(_w, email, functie));
-			else if (string.IsNullOrWhiteSpace(email)) Assert.Throws<WerknemerException>(() => b.VoegWerknemerToeInBedrijf(_w, email, functie));
-			else Assert.Throws<WerknemerInfoException>(() => b.VoegWerknemerToeInBedrijf(_w, email, functie));
+			if (string.IsNullOrWhiteSpace(functie))
+				Assert.Throws<BedrijfException>(() => b.VoegWerknemerToeInBedrijf(_w, email, functie));
+			else if (string.IsNullOrWhiteSpace(email))
+				Assert.Throws<WerknemerException>(() => b.VoegWerknemerToeInBedrijf(_w, email, functie));
+			else
+				Assert.Throws<WerknemerInfoException>(() => b.VoegWerknemerToeInBedrijf(_w, email, functie));
 			//"Werknemer - VoegBedrijfEnFunctieToeAanWerknemer - werknemer is in dit bedrijf al werkzaam onder deze functie"
 			b.VoegWerknemerToeInBedrijf(_w, _e, _f1);
 			Assert.Throws<WerknemerException>(() => b.VoegWerknemerToeInBedrijf(_w, _e, _f1));
@@ -237,7 +258,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 
 		#region UnitTest Verwijder Werknemer
 		[Fact]
-		public void VerwijderWerknemerUitBedrijf_Valid() {
+		public void VerwijderWerknemerUitBedrijf_Valid()
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.VoegWerknemerToeInBedrijf(_w, _e, _f1);
 			b.VerwijderWerknemerUitBedrijf(_w);
@@ -245,7 +267,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		}
 
 		[Fact]
-		public void VerwijderWerknemerUitBedrijf_Invalid() {
+		public void VerwijderWerknemerUitBedrijf_Invalid()
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 
 			Assert.Throws<BedrijfException>(() => b.VerwijderWerknemerUitBedrijf(null));
@@ -255,7 +278,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 
 		#region UnitTest Geef Werknemers
 		[Fact]
-		public void GeefWerknemers_Valid() {
+		public void GeefWerknemers_Valid()
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			b.VoegWerknemerToeInBedrijf(_w, _e, _f1);
 			Assert.Equal(_w, b.GeefWerknemers()[0]);
@@ -264,7 +288,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 
 		#region UnitTest Bedrijf is gelijk
 		[Fact]
-		public void BedrijfIsGelijk_Valid() {
+		public void BedrijfIsGelijk_Valid()
+		{
 			Bedrijf b = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Assert.True(b.BedrijfIsGelijk(b));
 		}
@@ -276,7 +301,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData(10, "bedrijf", "BE0676747521", "876543210", "bedrijf@email.com", "bedrijfstraat 10")]
 		[InlineData(10, "bedrijf", "BE0676747521", "012345678", "anderbedrijf@email.com", "bedrijfstraat 10")]
 		[InlineData(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "anderebedrijfstraat 10")]
-		public void BedrijfIsGelijk_Invalid(uint id, string naam, string btwNummer, string telefoonNummer, string email, string adres) {
+		public void BedrijfIsGelijk_Invalid(long id, string naam, string btwNummer, string telefoonNummer, string email, string adres)
+		{
 			Bedrijf b1 = new(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10");
 			Bedrijf b2 = new(id, naam, btwNummer, telefoonNummer, email, adres);
 			Assert.False(b1.BedrijfIsGelijk(b2));
@@ -301,7 +327,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 
 		[InlineData(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "     bedrijfstraat 10", 10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10")]
 		[InlineData(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10     ", 10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "bedrijfstraat 10")]
-		public void ctor_Valid(uint idIn, string naamIn, string btwNummerIn, string telefoonNummerIn, string emailIn, string adresIn, uint idUit, string naamUit, string btwNummerUit, string telefoonNummerUit, string emailUit, string adresUit) {
+		public void ctor_Valid(long idIn, string naamIn, string btwNummerIn, string telefoonNummerIn, string emailIn, string adresIn, long idUit, string naamUit, string btwNummerUit, string telefoonNummerUit, string emailUit, string adresUit)
+		{
 			Bedrijf b = new(idIn, naamIn, btwNummerIn, telefoonNummerIn, emailIn, adresIn);
 			Assert.Equal(idUit, b.Id);
 			Assert.Equal(naamUit, b.Naam);
@@ -365,7 +392,8 @@ namespace xUnitBezoekersRegistratiesysteem.Domeinen {
 		[InlineData(10, "bedrijf", "BE0676747521", "012345678", "bedrijf@email.com", "\v")]
 
 		[InlineData(10, null, null, null, null, null)]
-		public void ctor_Invalid(uint id, string naam, string btwNummer, string telefoonNummer, string email, string adres) {
+		public void ctor_Invalid(long id, string naam, string btwNummer, string telefoonNummer, string email, string adres)
+		{
 			Assert.Throws<BedrijfException>(() => new Bedrijf(id, naam, btwNummer, telefoonNummer, email, adres));
 		}
 		#endregion
